user_service/requirements.txt

flasgger==0.9.7.1
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mistune==3.1.2
packaging==24.2
pluggy==1.5.0
pytest==8.3.5
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
SQLAlchemy==2.0.39
typing_extensions==4.12.2
Werkzeug==3.1.3
PyMySQL==1.1.1
rabbitmq==0.2.0
redis==5.2.1
celery==5.2.7
python-consul==1.1.0     # Consul 服务发现集成
cryptography==41.0.3